+++ input_validation/module_utils/common_utils/data_validation.py
@@ -1,10 +1,10 @@
     # Main L1 Validation code. Get the JSON schema and input file to validate
 
 import json
-import ansible.module_utils.common.data_fetch as get
-import ansible.module_utils.common.en_us_validation_msg
+import ansible.module_utils.common_utils.data_fetch as get
+import ansible.module_utils.common_utils.en_us_validation_msg
 import jsonschema
-import ansible.module_utils.common.logical_validation
+import ansible.module_utils.common_utils.logical_validation
 
 
 def schema(input_file_path, schema_file_path, passwords_set, omnia_base_dir, project_name, logger, module):
