+++ input_validation/module_utils/validation_flows/common_validation.py
@@ -13,9 +13,9 @@
 # limitations under the License.
 
 import json
-import ansible.module_utils.common.validation_utils
-from ansible.module_utils.common import config
-import ansible.module_utils.common.en_us_validation_msg
+import ansible.module_utils.common_utils.validation_utils
+from ansible.module_utils.common_utils import config
+import ansible.module_utils.common_utils.en_us_validation_msg
 
 file_names = config.files
 create_error_msg = validation_utils.create_error_msg
