+++ input_validation/module_utils/validation_flows/provision_validation.py
@@ -13,9 +13,9 @@
 # limitations under the License.
 
 import json
-import ansible.module_utils.common.validation_utils
-import ansible.module_utils.common.config
-import ansible.module_utils.common.en_us_validation_msg
+import ansible.module_utils.common_utils.validation_utils
+import ansible.module_utils.common_utils.config
+import ansible.module_utils.common_utils.en_us_validation_msg
 
 file_names = config.files
 create_error_msg = validation_utils.create_error_msg
